Working

+++ pi4j-api/src/main/java/com/pi4j/annotation/AddPwmPreset.java
@@ -35,5 +35,5 @@ import java.lang.annotation.*;
 public @interface AddPwmPreset {
     String name();
     int frequency() default -1;
-    int dutyCycle() default -1;
+    float dutyCycle() default -1;
 }
