+++ pi4j-api/src/main/java/com/pi4j/annotation/InitialValue.java
@@ -35,5 +35,5 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.FIELD})
 public @interface InitialValue {
-    int value();
+    double value();
 }
